moved to test bed branch
trweking on the make file got some stuff to work on latter.   for now
i'll be sated with art icon to pull request
with job inteviews on coming next week.   may get to the install
funtion/s when i can...  cut/paste a few from dhsell's make... into this
copy see if it copies files in vm.
adds symlink , adds a desktop , and icon